fix expectDoesNotThrow
- `expectDoesNotThrow` was `@autoclosure () throws -> T`
- `expectThrows      ` was `() throws -> T

but both were called with an explicit closure. That had the really bad
effect that `T` in `expectDoesNotThrow` just became the _function_ type
but obviously we were never actually running that function as its real
type as lost.

There are basically two ways out of this and I implemented them both:

- both functions should take closures that return `Void`, not `T`. The
  value is lost anyway (due to `XCTAssertNoThrow` just throwing it away).
- whether the functions should take `@autoclosure` or not is up to
  debate. Because they were all called with explicit closures I opted
  for no autoclosures.

Summary: `() -> T` + `@autoclosure`  + ignoring the T (as `XCTAssertNoThrow` does) = desaster
as `T` can become a thunk that produces the type that the programmer
expects `T` to be.